+++ fs/sul/id/proof.f.ts
@@ -1,4 +1,4 @@
-import { assert, assertEq } from '../../dev/module.f.ts'
+import { assert, assertEq } from '../../asserts/module.f.ts'
 import { mask } from '../../types/bigint/module.f.ts'
 import { vec } from '../../types/bit_vec/module.f.ts'
 import { compress, hashId, isHash, isRaw, level3Id, rawId } from './module.f.ts'
